fix: 모든 예금 상품 api 500 에러 수정
